requirements.txt
Flask
SQLAlchemy
Flask-Migrate
requests
python-dotenv
psycopg2-binary